+++ src/main/java/org/ays/emergency_application/model/response/EmergencyEvacuationApplicationResponse.java
@@ -3,7 +3,7 @@ package org.ays.emergency_application.model.response;
 import lombok.Getter;
 import lombok.Setter;
 import org.ays.common.model.AysPhoneNumber;
-import org.ays.emergency_application.model.entity.EmergencyEvacuationApplicationStatus;
+import org.ays.emergency_application.model.enums.EmergencyEvacuationApplicationStatus;
 
 import java.time.LocalDateTime;
